feat(data): Introduce Standalone API for NgRx Data
