+++ modules/data/src/entity-data-without-effects.module.ts
@@ -1,34 +1,10 @@
+import { ModuleWithProviders, NgModule } from '@angular/core';
+import { EntityDataModuleConfig } from './entity-data-config';
 import {
-  ModuleWithProviders,
-  NgModule,
-  Inject,
-  Injector,
-  InjectionToken,
-  Optional,
-  OnDestroy,
-} from '@angular/core';
-
-import {
-  Action,
-  combineReducers,
-  MetaReducer,
-  ReducerManager,
-  StoreModule,
-} from '@ngrx/store';
-
-import { EntityCache } from './reducers/entity-cache';
-import { EntityCacheReducerFactory } from './reducers/entity-cache-reducer';
-import {
-  ENTITY_CACHE_NAME,
-  ENTITY_CACHE_NAME_TOKEN,
-  ENTITY_CACHE_META_REDUCERS,
-  INITIAL_ENTITY_CACHE_STATE,
-} from './reducers/constants';
-
-import {
-  EntityDataModuleConfig,
-  _provideEntityDataWithoutEffects,
-} from './provide_entity_data';
+  provideRootEntityDataWithoutEffects,
+  ENTITY_DATA_WITHOUT_EFFECTS_PROVIDERS,
+  initializeEntityDataWithoutEffects,
+} from './provide-entity-data';
 
 /**
  * Module without effects or dataservices which means no HTTP calls
@@ -37,66 +13,19 @@ import {
  * therefore opt-out of @ngrx/effects for entities
  */
 @NgModule({
-  imports: [
-    StoreModule, // rely on Store feature providers rather than Store.forFeature()
-  ],
-  providers: [],
+  providers: [ENTITY_DATA_WITHOUT_EFFECTS_PROVIDERS],
 })
-export class EntityDataModuleWithoutEffects implements OnDestroy {
-  private entityCacheFeature: any;
-
+export class EntityDataModuleWithoutEffects {
   static forRoot(
     config: EntityDataModuleConfig
   ): ModuleWithProviders<EntityDataModuleWithoutEffects> {
     return {
       ngModule: EntityDataModuleWithoutEffects,
-      providers: [..._provideEntityDataWithoutEffects(config)],
+      providers: [provideRootEntityDataWithoutEffects(config)],
     };
   }
 
-  constructor(
-    private reducerManager: ReducerManager,
-    entityCacheReducerFactory: EntityCacheReducerFactory,
-    private injector: Injector,
-    // optional params
-    @Optional()
-    @Inject(ENTITY_CACHE_NAME_TOKEN)
-    private entityCacheName: string,
-    @Optional()
-    @Inject(INITIAL_ENTITY_CACHE_STATE)
-    private initialState: any,
-    @Optional()
-    @Inject(ENTITY_CACHE_META_REDUCERS)
-    private metaReducers: (
-      | MetaReducer<EntityCache, Action>
-      | InjectionToken<MetaReducer<EntityCache, Action>>
-    )[]
-  ) {
-    // Add the @ngrx/data feature to the Store's features
-    // as Store.forFeature does for StoreFeatureModule
-    const key = entityCacheName || ENTITY_CACHE_NAME;
-
-    initialState =
-      typeof initialState === 'function' ? initialState() : initialState;
-
-    const reducers: MetaReducer<EntityCache, Action>[] = (
-      metaReducers || []
-    ).map((mr) => {
-      return mr instanceof InjectionToken ? injector.get(mr) : mr;
-    });
-
-    this.entityCacheFeature = {
-      key,
-      reducers: entityCacheReducerFactory.create(),
-      reducerFactory: combineReducers,
-      initialState: initialState || {},
-      metaReducers: reducers,
-    };
-    reducerManager.addFeature(this.entityCacheFeature);
-  }
-
-  // eslint-disable-next-line @angular-eslint/contextual-lifecycle
-  ngOnDestroy() {
-    this.reducerManager.removeFeature(this.entityCacheFeature);
+  constructor() {
+    initializeEntityDataWithoutEffects();
   }
 }
