+++ modules/data/src/entity-data.module.ts
@@ -5,11 +5,14 @@ import { EffectsModule, EffectSources } from '@ngrx/effects';
 import { EntityCacheEffects } from './effects/entity-cache-effects';
 import { EntityEffects } from './effects/entity-effects';
 
-import { EntityDataModuleWithoutEffects } from './entity-data-without-effects.module';
+import { EntityDataModuleConfig } from './entity-data-config';
 import {
-  EntityDataModuleConfig,
-  _provideEntityData,
-} from './provide_entity_data';
+  ENTITY_DATA_PROVIDERS,
+  ENTITY_DATA_WITHOUT_EFFECTS_PROVIDERS,
+  initializeEntityData,
+  provideRootEntityData,
+  provideRootEntityDataWithoutEffects,
+} from './provide-entity-data';
 
 /**
  * entity-data main module includes effects and HTTP data services
@@ -17,10 +20,7 @@ import {
  * No `forFeature` yet.
  */
 @NgModule({
-  imports: [
-    EntityDataModuleWithoutEffects,
-    EffectsModule, // do not supply effects because can't replace later
-  ],
+  providers: [ENTITY_DATA_WITHOUT_EFFECTS_PROVIDERS, ENTITY_DATA_PROVIDERS],
 })
 export class EntityDataModule {
   static forRoot(
@@ -28,33 +28,14 @@ export class EntityDataModule {
   ): ModuleWithProviders<EntityDataModule> {
     return {
       ngModule: EntityDataModule,
-      providers: [..._provideEntityData(config)],
+      providers: [
+        provideRootEntityDataWithoutEffects(config),
+        provideRootEntityData(config),
+      ],
     };
   }
 
-  constructor(
-    private effectSources: EffectSources,
-    entityCacheEffects: EntityCacheEffects,
-    entityEffects: EntityEffects
-  ) {
-    // We can't use `forFeature()` because, if we did, the developer could not
-    // replace the entity-data `EntityEffects` with a custom alternative.
-    // Replacing that class is an extensibility point we need.
-    //
-    // The FEATURE_EFFECTS token is not exposed, so can't use that technique.
-    // Warning: this alternative approach relies on an undocumented API
-    // to add effect directly rather than through `forFeature()`.
-    // The danger is that EffectsModule.forFeature evolves and we no longer perform a crucial step.
-    this.addEffects(entityCacheEffects);
-    this.addEffects(entityEffects);
-  }
-
-  /**
-   * Add another class instance that contains effects.
-   * @param effectSourceInstance a class instance that implements effects.
-   * Warning: undocumented @ngrx/effects API
-   */
-  addEffects(effectSourceInstance: any) {
-    this.effectSources.addEffects(effectSourceInstance);
+  constructor() {
+    initializeEntityData();
   }
 }
